python/tts_services.py
"""
TTS Services - Text-to-Speech implementations
"""

from typing import Optional
import google.generativeai as genai

from config import settings


class GoogleTTSService:
    """Google TTS Service using Gemini"""

    def __init__(self, api_key: str):
        genai.configure(api_key=api_key)
        self.model = genai.GenerativeModel('gemini-1.0-pro')

    async def generate_audio(self, text: str) -> bytes:
        """Generate audio from text"""
        # Note: This is a placeholder. Actual implementation would depend
        # on the specific TTS API being used with Gemini
        # For now, we'll return empty bytes
        return b""


class GoogleSTTService:
    """Google STT Service"""

    def __init__(self, api_key: str):
        self.api_key = api_key

    async def transcribe_audio(self, audio_data: bytes) -> str:
        """Transcribe audio to text"""
        # Placeholder implementation
        return ""